+++ apps/docs/src/content/03-components/structure/column-layout/examples/hidden-columns.tsx
@@ -1,0 +1,28 @@
+import {
+  ColumnLayout,
+  Text,
+  Image,
+  Section,
+  Heading,
+} from "@mittwald/flow-react-components";
+
+<ColumnLayout s={[1, null]} m={[2, 1]} l={[5, 1]}>
+  <Section>
+    <Heading>Lorem ipsum dolor sit amet</Heading>
+    <Text>
+      Lorem ipsum dolor sit amet consectetur adipisicing
+      elit. Cumque eius quam quas vel voluptas, ullam
+      aliquid fugit. Voluptate harum accusantium rerum ullam
+      modi blanditiis vitae, laborum ea tempore, dolore
+      voluptas. Earum pariatur, similique corrupti id
+      officia perferendis. Labore, similique. Earum, quas
+      in. At dolorem corrupti blanditiis nulla deserunt
+      laborum! Corrupti delectus aspernatur nihil nulla
+      obcaecati ipsam porro sequi rem? Quam.
+    </Text>
+  </Section>
+  <Image
+    src="https://mittwald.github.io/flow/assets/mittwald_logo_rgb.jpg"
+    alt="mittwald"
+  />
+</ColumnLayout>;