+++ packages/components/src/components/ColumnLayout/lib/getColumns.ts
@@ -1,9 +1,7 @@
-export const getColumns = (values: number[]): string => {
-  if (values.length < 1) {
-    throw new Error("Column layout array is empty");
-  }
-
-  const fractionValues = values.map((value) => `${value}fr`);
+export const getColumns = (values: (number | null)[]): string => {
+  const fractionValues = values
+    .filter((v) => v !== null)
+    .map((value) => `${value}fr`);
 
   return fractionValues.join(" ");
 };